+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-### Version 1.4.0
+### Version 1.5.0
+
 12-Oct-2018
 
-- Gracefully handle `null`, `undefined`, or non-`string` values passed as phone number args.
+- Recognize toll-free numbers with an NXX ending in `11` as valid. Non-toll-free numbers
+  with an NXX ending in `11` are not valid.
+
+### Version 1.4.0
+
+12-Oct-2018
+
+- Gracefully handle `null`, `undefined`, or non-`string` values passed as phone number
+  args.